requirements.txt
auraloss
h5py
matplotlib
numpy
pyyaml
scipy
setuptools
torch
torchaudio
tensorboard